IDEA-202111 - Refactor code to remove code duplicates and make elements less visible
